% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuc_chrII-data.R
\docType{data}
\name{nuc_chrII}
\alias{nuc_chrII}
\title{Sample chromosome II nucleosome calls}
\format{An object of class list, containing two GRanges objects,
one for each experiment}
\usage{
data(nuc_chrII)
}
\value{
Nucleosome positions in S. cerevisiae chrII:1-400000 for G2 and M
}
\description{
Nucleosome positions in S. cerevisiae (sacCer3) chrII, derived from
MNase-seq data, for cell cycle phases G2 and M
}
\references{
Deniz, 0., Flores, O., et al. (2016) Scientific Reports 6, 19729
(\href{https://www.ebi.ac.uk/ena/data/view/PRJEB6970}{ENA: PRJEB6970})
}
\keyword{datasets}
