% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucDyn-class.R
\name{set.a}
\alias{set.a}
\alias{set.b}
\title{Accessors to a NucDyn object}
\usage{
set.a(x)

set.b(x)
}
\arguments{
\item{x}{a NucDyn object}
}
\value{
\itemize{
\item {set.a} \code{GRangesList} with the reads paired from ref1
\item {set.b} \code{GRangesList} with the reads paired from ref2
}
}
\description{
Accessors to a NucDyn object
}
