% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucDyn-package.R
\docType{package}
\name{NucDyn-package}
\alias{NucDyn-package}
\title{Comparative dynamics analysis at a read level bewteen NGS nucleosome
coverage maps.}
\description{
Nucleosome Dynamics allows detection of changes in the positions of
nucleosomes at single read level between two reference nucleosome maps.
Changes include shifts (moves), indels (insertions/deletions) of the
existing reads in initial state to match the reference final state.
}
\details{
\tabular{ll}{
Package: \tab NucDyn       \cr
Type:    \tab Package      \cr
Version: \tab 0.1          \cr
Date:    \tab 2014-08-19   \cr
License: \tab BSD_2_clause \cr
}
}
\author{
Ricard Illa, Maintainer: Diana Buitrago \href{mailto:diana.buitrago@irbbarcelona.org}{diana.buitrago@irbbarcelona.org}
}
