% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsG2_chrII-data.R
\docType{data}
\name{readsG2_chrII}
\alias{readsG2_chrII}
\title{Sample reads from MNase-seq G2 cells in yeast}
\format{An object of class RangedData}
\usage{
data(readsG2_chrII)
}
\value{
MNase-seq mapped reads for S. cerevisiae chrII:1-400000 in G2
}
\description{
Mapped reads from an MNaseSeq experiment for S. cerevisiae (sacCer3) in
G2 cell cycle phase
}
\references{
Deniz, 0., Flores, O., et al. (2016) Scientific Reports 6, 19729
(\href{https://www.ebi.ac.uk/ena/data/view/PRJEB6970}{ENA: PRJEB6970})
}
\keyword{datasets}
