% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsM_chrII-data.R
\docType{data}
\name{readsM_chrII}
\alias{readsM_chrII}
\title{Sample reads from MNase-seq M cells in yeast}
\format{An object of class RangedData}
\usage{
data(readsM_chrII)
}
\value{
MNase-seq mapped reads for S. cerevisiae chrII:1-400000 in M
}
\description{
Mapped reads from an MNaseSeq experiment for S. cerevisiae (sacCer3) in
M cell cycle phase
}
\references{
Deniz, 0., Flores, O., et al. (2016) Scientific Reports 6, 19729
(\href{https://www.ebi.ac.uk/ena/data/view/PRJEB6970}{ENA: PRJEB6970})
}
\keyword{datasets}
