% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\docType{data}
\name{sample_chrII}
\alias{sample_chrII}
\title{Sample chromosome II coverage data}
\format{An object of class list}
\usage{
data(sample_chrII)
}
\description{
Data from a MNaseSeq experiment in yeast in the following mitosis phases:
G2 and M
}
\keyword{datasets}
