% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHotspots.R
\name{applyThreshold}
\alias{applyThreshold}
\title{Apply threshold}
\usage{
applyThreshold(hs, indel.thresh, shift.thresh, indel.nreads, shift.nreads)
}
\arguments{
\item{hs}{Hotspots returned by findHotspots.}

\item{indel.thresh}{threshold for the indels.}

\item{shift.thresh}{threshold for the shifts.}

\item{indel.nreads}{minimum number of reads for the indels.}

\item{shift.nreads}{minimum number of reads for the shifts.}
}
\value{
a hotspots \code{data.frame} filered by the thresholds.
}
\description{
Apply an indel threshold and a shift threshold to the hotspots
}
\author{
Ricard Illa,
Diana Buitrago \email{diana.buitrago@irbbarcelona.org}
}
\keyword{manip}
