% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHotspots.R
\docType{methods}
\name{findHotspots}
\alias{findHotspots}
\alias{findHotspots,NucDyn-method}
\title{Find hotspots in a NucDyn object.}
\usage{
findHotspots(dyn, wins = 10000, indel.threshold = NULL,
  shift.threshold = NULL, mc.cores = 1)

\S4method{findHotspots}{NucDyn}(dyn, wins = 10000,
  indel.threshold = NULL, shift.threshold = NULL, mc.cores = 1)
}
\arguments{
\item{dyn}{NucDyn object with the dynamic to analyze.}

\item{wins}{Size of the window in base-pairs where the relative scores are
computed}

\item{indel.threshold}{Maximum p-value for an insertion or delition hotspot
to be considered significant.}

\item{shift.threshold}{Maximum p-value for shift hotspot to be considered
significant.}

\item{mc.cores}{If \code{parallel} support, the number of cores available. This
option is only used if the provided sets are from more than one
chromosome.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item chrom: Chromosome name.
\item coord: Genomic coordinates (average dyad position of affected
nucleosomes).
\item type: The type of the hotspot (as listed above).
\item nreads: Number of reads involved in the hotspot.
}
}
\description{
Find hotspots from a given nucleosome dynamics.
}
\author{
Ricard Illa \email{ricard.illa@irbbarcelona.org}
}
\keyword{manip}
