\name{findPVals}
\alias{findPVals}
\title{
    Calculate a vector of p-values expressing the difference between two
    coverages.
}
\description{
    Calculate a vector of p-values expressing the difference between two
    coverages. Works by windows.
}
\arguments{
    \item{x}{
        First coverage.
    }
    \item{y}{
        Second coverage.
    }
    \item{wins}{
        Size of the window.
    }
}
\author{
    Ricard Illa \email{ricard.illa@irbbarcelona.org}
}
\keyword{manip}
