% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucDyn-class.R
\docType{methods}
\name{nucdyn-accessors}
\alias{nucdyn-accessors}
\alias{set.a}
\alias{set.b}
\alias{set.a,NucDyn-method}
\alias{set.b,NucDyn-method}
\title{Accessors to a NucDyn object}
\usage{
set.a(x)

set.b(x)

\S4method{set.a}{NucDyn}(x)

\S4method{set.b}{NucDyn}(x)
}
\arguments{
\item{x}{a NucDyn object}
}
\value{
\itemize{
\item {set.a} \code{GRangesList} with the reads paired from ref1
\item {set.b} \code{GRangesList} with the reads paired from ref2
}
}
\description{
Accessors to a NucDyn object
}
\examples{
    data(readsG2_chrII)
    data(readsM_chrII)
    \donttest{dyn <- nucleosomeDynamics(setA=readsG2_chrII, setB=readsM_chrII)}
    \donttest{set.a(dyn)}
    \donttest{set.b(dyn)}

}
