% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chrII-data.R
\docType{data}
\name{sample_chrII}
\alias{sample_chrII}
\title{Sample chromosome II coverage data}
\format{An object of class list}
\usage{
data(sample_chrII)
}
\value{
Nucleosome coverage in S. cerevisiae chrII for G2 and M
}
\description{
Data from a MNaseSeq experiment in S. cerevisiae (sacCer3) in the
following cell cycle phases: G2 and M
}
\references{
Deniz, 0., Flores, O., et al. (2016) Scientific Reports 6, 19729
(\href{https://www.ebi.ac.uk/ena/data/view/PRJEB6970}{ENA: PRJEB6970})
}
\keyword{datasets}
